+++ mm2src/mm2_main/src/rpc/lp_commands/mod.rs
@@ -1,4 +1,3 @@
-pub(crate) mod crypto_ctx;
 pub(crate) mod db_id;
 pub mod legacy;
 pub(crate) mod one_inch;
